requirements.txt
firebase_admin==6.6.0
Flask==3.1.0
google-generativeai==0.8.4
kaleido==0.2.1
loguru==0.7.3
numpy==2.2.2
openai==1.60.1
plotly==6.0.0
requests==2.32.3
pre-commit==4.1.0
pyTelegramBotAPI==4.26.0
pytest==8.3.4
pyyaml==6.0.2
ruff==0.9.3